requirements.txt
beautifulsoup4
urlextract
python-magic
imagehash
pillow